mcp-agents/gdrive-pdf-upload-mcp-agent/requirements.txt
# Google Drive PDF Upload MCP Agent
mcp>=1.0.0
fastmcp>=2.0.0
uagents>=0.1.0
uagents-adapter[mcp]>=0.1.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
python-dotenv>=1.0.0
